JDBC 드라이버 추가 및 환경설정 MariaDB 접속정보 추가

+++ src/main/java/com/example/pcroom/Repository/ReservationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+public interface ReservationRepository extends JpaRepository<Reservation, Integer> {
 }
